Move sgx/ to core/
