+++ GNUmakefile
@@ -15,7 +15,7 @@ DOCKER_TAG ?= latest
 SGX_ENABLED ?= no
 SGX_SIMULATION ?= yes
 SGX_ENCLAVE := enclave.signed.so
-SGX_TARGET := ./sgx/target/$(ENVIRONMENT)/
+SGX_TARGET := ./core/sgx/target/$(ENVIRONMENT)/
 
 ifneq (,$(filter yes true,$(SGX_ENABLED)))
 	GOFLAGS += -tags=sgx_enclave
@@ -79,8 +79,8 @@ dockerpush: ## Push the docker image to dockerhub
 
 .PHONY: $(SGX_ENCLAVE)
 $(SGX_ENCLAVE):
-	@ENVIRONMENT=$(ENVIRONMENT) SGX_ENABLED=$(SGX_ENABLED) SGX_SIMULATION=$(SGX_SIMULATION) make -C sgx/
-	@ln -f $(SGX_TARGET)/libadapters.so sgx/target/libadapters.so
+	@ENVIRONMENT=$(ENVIRONMENT) SGX_ENABLED=$(SGX_ENABLED) SGX_SIMULATION=$(SGX_SIMULATION) make -C core/sgx/
+	@ln -f $(SGX_TARGET)/libadapters.so core/sgx/target/libadapters.so
 
 .PHONY: enclave
 enclave: $(SGX_ENCLAVE)
