+++ core/sgx/bigdecimal-rs/COPYRIGHT
