+++ core/sgx/bigdecimal-rs/examples/repeated_squares.rs
