+++ core/sgx/bigdecimal-rs/src/macros.rs
