+++ core/sgx/enclave/enclave.lds
