+++ core/sgx/enclave/src/result.rs
