+++ core/sgx/libadapters/.gitignore
