+++ core/sgx/libadapters/adapters.h
