+++ core/sgx/libadapters/src/lib.rs
