+++ core/sgx/libadapters/src/multiply.rs
