+++ tools/ci/rust_test
@@ -2,6 +2,6 @@
 
 set -e
 
-pushd sgx/utils 2>/dev/null
+pushd core/sgx/utils 2>/dev/null
 cargo test
 popd 2>/dev/null
